Comments api and adjust delete routes

+++ server/src/routes/snippets.ts
@@ -21,6 +21,7 @@ router.get("/", async (req: Request, res: Response) => {
   const amount = req.query.amount ? parseInt(req.query.amount as string) : 10;
   const page = req.query.page ? parseInt(req.query.page as string) : 1;
   const offset = (page - 1) * amount;
+  if (page <= 0 || amount <= 0) return res.sendStatus(400);
 
   try {
     const result = await pool.query(
@@ -38,7 +39,7 @@ router.get("/", async (req: Request, res: Response) => {
 });
 
 /**
- * @api {post} /api/snippets/:id Get a snippet
+ * @api {post} /api/snippets/:id Get a single snippet
  */
 router.get("/:id", async (req: Request, res: Response) => {
   const { id } = req.params;
@@ -59,12 +60,14 @@ router.get("/:id", async (req: Request, res: Response) => {
 
 /**
  * @api {post} /api/snippets/ Create a new snippet
+ * @param {string} [title] the title of the snippet
+ * @param {string} [content] the content of the snippet
  */
 router.post("/", validateToken, async (req: Request, res: Response) => {
   const { userId } = req;
   const { title, content } = req.body;
 
-  if (!title || !content) {
+  if (!title || title.length === 0 || !content || content.length === 0) {
     return res.sendStatus(400);
   }
 
@@ -84,6 +87,8 @@ router.post("/", validateToken, async (req: Request, res: Response) => {
 
 /**
  * @api {patch} /api/snippets/:id Edit a snippet
+ * @param {string} [title] New title
+ * @param {string} [content] New content
  */
 router.put("/:id", validateToken, async (req: Request, res: Response) => {
   const { userId } = req;
@@ -117,11 +122,17 @@ router.delete("/:id", validateToken, async (req: Request, res: Response) => {
   const { id } = req.params;
 
   try {
+    const result = await pool.query(
+      "SELECT * FROM snippets WHERE id = $1 AND user_id = $2",
+      [id, userId]
+    );
+    if (result.rowCount === 0) return res.sendStatus(404);
+
     await pool.query("DELETE FROM snippets WHERE id = $1 AND user_id = $2", [
       id,
       userId,
     ]);
-    return res.sendStatus(200);
+    return res.sendStatus(204);
   } catch (err) {
     console.error(err);
     return res.sendStatus(500);
